Added regression test for Content-Security policy builder - no more unnecessary blank space characters added
